requirements.txt
flask==2.3.3
anthropic==0.40.0
python-dotenv==1.0.0
GitPython==3.1.40
python-dateutil==2.8.2